requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
anyio==3.7.1
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.0
apache-airflow-providers-ftp==3.5.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.3.0
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backoff==1.10.0
beautifulsoup4==4.13.3
blinker==1.6.2
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
confluent-kafka==2.10.0
connexion==2.14.2
contourpy==1.3.1
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.3
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-useragent==2.1.0
Flask==2.2.5
Flask-AppBuilder==4.3.3
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fonttools==4.57.0
frozenlist==1.4.0
google-re2==1.1
googleapis-common-protos==1.70.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==6.0.1
importlib_metadata==8.7.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.8
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.10.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
numpy==2.2.4
openpyxl==3.1.5
opentelemetry-api==1.15.0
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-semantic-conventions==0.36b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.3
pathspec==0.11.2
pendulum==2.1.2
pillow==11.1.0
pipdeptree==2.23.1
pluggy==1.2.0
prison==0.2.1
protobuf==4.25.0
psutil==5.9.5
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyarrow==19.0.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.2.3
pyspark==3.5.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.5.2
rich-argparse==1.2.0
rpds-py==0.9.2
seaborn==0.13.2
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
squarify==0.4.4
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2025.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
vnai==2.0.4
vnstock==3.2.4
vnstock_ezchart==0.0.2
Werkzeug==2.2.3
wordcloud==1.9.4
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.21.0
